kraken-std/: breaking change: Move Kraken task definitions in `kraken.std.docker` into `kraken.std.docker.tasks` module
