+++ kraken-std/src/kraken/std/docker/tasks/docker_build_task.py
@@ -8,12 +8,12 @@ from pathlib import Path
 from kraken.common import flatten, not_none
 from kraken.core import Project, Property, TaskStatus
 
-from kraken.std.docker.util import update_run_commands
+from kraken.std.docker.util.dockerfile import update_run_commands
 
-from . import DockerBuildTask
+from .base_build_task import BaseBuildTask
 
 
-class NativeBuildTask(DockerBuildTask):
+class NativeBuildTask(BaseBuildTask):
     """Implements building a Docker image using the native `docker build` command."""
 
     #: Whether to use Docker Buildkit. Enabled by default.
@@ -23,13 +23,13 @@ class NativeBuildTask(DockerBuildTask):
         super().__init__(name, project)
         self.preprocess_dockerfile.set(True)
 
-    # DockerBuildTask
+    # BaseBuildTask overrides
 
     def _preprocess_dockerfile(self, dockerfile: Path) -> str:
         mount_string = " ".join(f"--mount=type=secret,id={sec}" for sec in self.secrets.get().keys()) + " "
         return update_run_commands(dockerfile.read_text(), prefix=mount_string)
 
-    # Task
+    # Task overrides
 
     def finalize(self) -> None:
         if self.cache_repo.get() is not None:
